Added functionality for y_servo & Added Comments
Added the functionality for Y_servo with calculation about difference
in date between the current date and the longest day of the year.
The logic works fine for the time when the y_Servo is increasing i.e.
difference between dates if positive.
Need to work on the part when the days are decreasing.
Also, added comments to understand the code better.