Client Panel repaired
The latest voltage and current data can now be fetched from the
database and seen. (Only the latest value can be seen). This values
also update every 10 seconds.

additional file of fetch_info.php is made for fetching the latest data.

The logic and code for “Click” to “Download” the arduino code is now
working along with the correct credentials been update from the
database to the write_file.php. A few corrections in the write_file.php
are made in order to have a correct exact arduino code as required.